Implement grid submission
Grid submission is quite naive: the game state simply keeps track of the solution reached by the solver and the game just compares each cell with each cell of the solution.
Note that in case the grid has multiple solutions (hopefully it never happens), the submission only accepts one of them, i.e. the one obtained by the solver